#24 Fixed cascade problem and update readme.md.

+++ src/main/java/cn/ncepu/voluntize/controller/department/Manage.java
@@ -15,6 +15,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
 
 @RestController("/department/service")
@@ -28,7 +29,7 @@ public class Manage extends BaseController {
 
     /**
      * 测试信息
-·    *
+     * ·    *
      */
     @RequestMapping("/saveActivity")
     @ResponseBody
@@ -50,17 +51,22 @@ public class Manage extends BaseController {
     }
 
     @RequestMapping("/cancel")
-    public void cancel(String activityId) {
-        activityService.deleteActivity(activityId);
+    public void cancel(String activityId, String stationId, String periodId) {
+        if (activityId != null)
+            activityService.deleteActivity(activityId);
+        if (stationId != null)
+            activityService.deleteActivityStation(stationId);
+        if (periodId != null)
+            activityService.deleteActivityPeriod(periodId);
     }
 
     @RequestMapping("/approve")
-    public void approve(@RequestBody ArrayList<String> recordId) {
+    public void approve(@RequestBody List<String> recordId) {
         participateService.accept(recordId);
     }
 
     @RequestMapping("/evaluate")
-    public void evaluate(@RequestBody ArrayList<EvaluateVo> evaluateVos) {
+    public void evaluate(@RequestBody List<EvaluateVo> evaluateVos) {
         participateService.evaluate(evaluateVos);
     }
 }
